backend/requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
sentence-transformers
